[SERVER] feat: Dockerfile update

See merge request s07-final/S07P31A103!12

+++ backend/portfolio/src/main/java/com/folaroid/portfolio/db/entity/User.java
@@ -10,7 +10,7 @@ public class User {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer userNo;
+    private long userNo;
 
     @Column(length = 60)
     private String userGithubId;
